+++ retail-modules/retail-colonel/retail-colonel-common/src/main/java/com.retail.colonel/domain/WarehouseEntity.java
@@ -1,5 +1,6 @@
 package com.retail.colonel.domain;
 
+import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 
@@ -43,6 +44,6 @@ public class WarehouseEntity implements Serializable {
 	/**
 	 * 状态
 	 */
-	private Long status;
+	private Integer status;
 
 }
