+++ retail-modules/retail-order/retail-order-common/src/main/java/com.retail.order/domain/ExpressageEntity.java
@@ -1,5 +1,6 @@
 package com.retail.order.domain;
 
+import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 
@@ -35,6 +36,6 @@ public class ExpressageEntity implements Serializable {
 	/**
 	 * 状态
 	 */
-	private Long status;
+	private Integer status;
 
 }
