+++ retail-modules/retail-order/retail-order-common/src/main/java/com.retail.order/domain/ReatilOrderInfoEntity.java
@@ -1,5 +1,6 @@
 package com.retail.order.domain;
 
+import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 
@@ -32,11 +33,11 @@ public class ReatilOrderInfoEntity implements Serializable {
 	/**
 	 * 订单id
 	 */
-	private Integer orderSn;
+	private Long orderSn;
 	/**
 	 * 商品spuId
 	 */
-	private Integer spuId;
+	private Long spuId;
 	/**
 	 * 商品总数量
 	 */
